+++ test/unit/events.js
@@ -4,7 +4,7 @@
 (function($){
 
   module('Events', { teardown: function teardown(){
-    $('.jquery-reel').unbind('loaded store recall').trigger('teardown');
+    $('.jquery-reel').unbind('loaded store recall openingDone play').trigger('teardown');
   }});
 
   asyncTest( 'Internal data setting triggers "store" event and passes name and value to the handler', function(){
@@ -50,6 +50,44 @@
           compare= false;
           start();
         }
+      })
+  });
+
+  asyncTest( '`"openingDone"` event is triggered at the end of opening animation', function(){
+    expect(2);
+    var
+      $reel= $('#image').reel({
+        opening: 1
+      })
+
+    $reel
+      .bind('openingDone', function(){
+        ok( true, '`"openingDone"` has been triggered');
+      })
+      .bind('play', function(){
+        ok( true, '`"play" event has fired closely following the `"openingDone"`');
+        start();
+      })
+  });
+
+  asyncTest( 'Default handling of `"openingDone"` event is cancelable by returning false from the handler', function(){
+    expect(2);
+    var
+      $reel= $('#image').reel({
+        opening: 1
+      })
+
+    $reel
+      .bind('openingDone', function(){
+        ok( true, '`"openingDone"` has been triggered and set to prevent event bubbling');
+        setTimeout(function(){
+          ok( true, 'The cancelled `"play"` correctly didn\'t fire (waited for it half a second)');
+          start();
+        }, 500);
+        return false
+      })
+      .bind('play', function(){
+        start();
       })
   });